Fixed compiler warning

+++ src/generator_convolution_forward_avx512.c
@@ -669,8 +669,8 @@ void libxsmm_generator_convolution_forward_avx512_c3_bf16( libxsmm_generated_cod
   step_size = 1;
 
   for (pi=0; pi<32; pi+=2) {
-    perm_array[pi] = pi/2;
-    perm_array[pi+1] = 16+pi/2;
+    perm_array[pi] = ((unsigned short) pi)/2;
+    perm_array[pi+1] = 16+ ((unsigned short) pi)/2;
   }
 
   for ( l_kw_unroll = 0; l_kw_unroll < i_kw_unroll; l_kw_unroll++ ) {
